refactor(ln-dlc-node): Move estimate_fee method to wallet module
