+++ crates/ln-dlc-node/src/node/wallet.rs
@@ -8,11 +8,14 @@ use crate::storage::TenTenOneStorage;
 use anyhow::Context;
 use anyhow::Result;
 use bdk_esplora::EsploraAsyncExt;
+use bitcoin::address::NetworkUnchecked;
 use bitcoin::secp256k1::SecretKey;
 use bitcoin::Address;
+use bitcoin::Amount;
 use bitcoin::OutPoint;
 use bitcoin::ScriptBuf;
 use bitcoin::TxOut;
+use lightning::chain::chaininterface::ConfirmationTarget;
 use std::sync::Arc;
 use tokio::task::spawn_blocking;
 
@@ -58,6 +61,19 @@ impl<D: BdkStorage, S: TenTenOneStorage, N: Storage + Send + Sync + 'static> Nod
 
     pub fn is_mine(&self, script_pubkey: &ScriptBuf) -> bool {
         self.wallet.is_mine(script_pubkey)
+    }
+
+    /// Estimate the fee for sending the given `amount_sats` to the given `address` on-chain with
+    /// the given `fee`.
+    pub fn estimate_fee(
+        &self,
+        address: Address<NetworkUnchecked>,
+        amount_sats: u64,
+        fee: ConfirmationTarget,
+    ) -> Result<Amount> {
+        let address = address.require_network(self.network)?;
+
+        self.wallet.estimate_fee(&address, amount_sats, fee)
     }
 
     /// Sync the state of the on-chain wallet against the blockchain.
